+++ src/main/java/de/canitzp/feederhelmet/data/FHRecipeProvider.java
@@ -1,17 +1,23 @@
 package de.canitzp.feederhelmet.data;
 
 import de.canitzp.feederhelmet.FeederHelmet;
+import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
-import net.minecraft.data.recipes.*;
+import net.minecraft.data.recipes.RecipeCategory;
+import net.minecraft.data.recipes.RecipeOutput;
+import net.minecraft.data.recipes.RecipeProvider;
+import net.minecraft.data.recipes.ShapedRecipeBuilder;
 import net.minecraft.world.item.Items;
 import net.neoforged.neoforge.common.Tags;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.concurrent.CompletableFuture;
+
 
 public class FHRecipeProvider extends RecipeProvider {
 
-    public FHRecipeProvider(PackOutput output) {
-        super(output);
+    public FHRecipeProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider) {
+        super(output, lookupProvider);
     }
 
     @Override