+++ src/main/java/de/canitzp/feederhelmet/item/ItemPhotosynthesisModule.java
@@ -1,6 +1,5 @@
 package de.canitzp.feederhelmet.item;
 
-import de.canitzp.feederhelmet.FeederHelmet;
 import de.canitzp.feederhelmet.data.localization.FHLocalizationKeys;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
@@ -8,9 +7,7 @@ import net.minecraft.network.chat.Style;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.level.Level;
 
-import javax.annotation.Nullable;
 import java.util.List;
 
 public class ItemPhotosynthesisModule extends Item {
@@ -20,8 +17,8 @@ public class ItemPhotosynthesisModule extends Item {
     }
     
     @Override
-    public void appendHoverText(ItemStack stack, @Nullable Level worldIn, List<Component> tooltip, TooltipFlag flagIn) {
-        super.appendHoverText(stack, worldIn, tooltip, flagIn);
+    public void appendHoverText(ItemStack stack, Item.TooltipContext context, List<Component> tooltip, TooltipFlag flagIn) {
+        super.appendHoverText(stack, context, tooltip, flagIn);
         tooltip.add(Component.translatable(FHLocalizationKeys.MODULE_PHOTOSYNTHESIS_DESCRIPTION).setStyle(Style.EMPTY.applyFormats(ChatFormatting.GRAY)));
     }
 }