+++ src/main/java/de/canitzp/feederhelmet/recipe/RecipeModuleAddition.java
@@ -1,12 +1,13 @@
 package de.canitzp.feederhelmet.recipe;
 
 import com.mojang.serialization.Codec;
+import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import de.canitzp.feederhelmet.FeederHelmet;
-import de.canitzp.feederhelmet.NBTHelper;
-import net.minecraft.core.RegistryAccess;
+import net.minecraft.core.HolderLookup;
 import net.minecraft.core.registries.BuiltInRegistries;
-import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.network.RegistryFriendlyByteBuf;
+import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.Container;
 import net.minecraft.world.item.Item;
@@ -15,18 +16,11 @@ import net.minecraft.world.item.crafting.RecipeSerializer;
 import net.minecraft.world.item.crafting.SmithingRecipe;
 import net.minecraft.world.level.Level;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
-public class RecipeModuleAddition implements SmithingRecipe {
+public record RecipeModuleAddition(Item helmet, String module, ItemStack outputStack) implements SmithingRecipe {
 
-    private final Item helmet;
-    private final String module;
-    private final ItemStack outputStack;
-
-    public RecipeModuleAddition(Item helmet, String module, ItemStack outputStack) {
-        this.helmet = helmet;
-        this.module = module;
-        this.outputStack = outputStack;
+    public RecipeModuleAddition(ResourceLocation loc, String module, ItemStack outputStack) {
+        this(BuiltInRegistries.ITEM.get(loc), module, outputStack);
     }
 
     @Override
@@ -36,7 +30,7 @@ public class RecipeModuleAddition implements SmithingRecipe {
 
     @Override
     public boolean isBaseIngredient(ItemStack stack) {
-        return stack.is(helmet) && !NBTHelper.isModulePresent(module, stack);
+        return stack.is(this.helmet) && !FeederHelmet.hasModule(stack, this.module);
     }
 
     @Override
@@ -46,62 +40,64 @@ public class RecipeModuleAddition implements SmithingRecipe {
 
     @Override
     public boolean matches(Container container, Level level) {
-        return isTemplateIngredient(container.getItem(0)) && isBaseIngredient(container.getItem(1)) && isAdditionIngredient(container.getItem(2));
+        return this.isTemplateIngredient(container.getItem(0)) && this.isBaseIngredient(container.getItem(1)) && this.isAdditionIngredient(container.getItem(2));
     }
 
     @Override
-    public ItemStack assemble(Container container, RegistryAccess access) {
+    public ItemStack assemble(Container container, HolderLookup.Provider access) {
         ItemStack assembled = this.getResultItem(access).copy();
-        // copy old nbt to new stack
-        assembled.getOrCreateTag().merge(container.getItem(1).getOrCreateTag());
+
+        // copy all components
+        assembled.applyComponents(container.getItem(1).getComponents());
         // set module flag
-        NBTHelper.addModule(module, assembled);
+        FeederHelmet.addModule(assembled, this.module);
         return assembled;
     }
 
     @Override
-    public ItemStack getResultItem(RegistryAccess access) {
-        return outputStack;
+    public ItemStack getResultItem(HolderLookup.Provider access) {
+        return this.outputStack;
     }
 
     @Override
     public RecipeSerializer<?> getSerializer() {
-        return FeederHelmet.MODULE_ADDITION_SERIALIZER.get();
-    }
-
-    public Item getBase() {
-        return this.helmet;
-    }
-
-    public ItemStack getOutputStack() {
-        return this.outputStack;
+        return Serializer.INSTANCE;
     }
 
     public static class Serializer implements RecipeSerializer<RecipeModuleAddition> {
 
-        @Override
-        public @Nullable RecipeModuleAddition fromNetwork(FriendlyByteBuf buffer) {
+        public static final Serializer INSTANCE = new Serializer();
+        private static final MapCodec<RecipeModuleAddition> CODEC = RecordCodecBuilder.mapCodec(instance -> instance.group(
+                ResourceLocation.CODEC.fieldOf("helmet").forGetter(recipe -> BuiltInRegistries.ITEM.getKey(recipe.helmet)),
+                Codec.STRING.fieldOf("module").forGetter(RecipeModuleAddition::module),
+                ItemStack.CODEC.fieldOf("result").forGetter(RecipeModuleAddition::outputStack)
+        ).apply(instance, RecipeModuleAddition::new));
+
+        public static final StreamCodec<RegistryFriendlyByteBuf, RecipeModuleAddition> STREAM_CODEC = StreamCodec.of(
+                Serializer::toNetwork, Serializer::fromNetwork
+        );
+
+        public static RecipeModuleAddition fromNetwork(RegistryFriendlyByteBuf buffer) {
             Item helmet = BuiltInRegistries.ITEM.get(buffer.readResourceLocation());
             String module = buffer.readUtf();
-            ItemStack outputStack = buffer.readItem();
+            ItemStack outputStack = ItemStack.STREAM_CODEC.decode(buffer);
             return new RecipeModuleAddition(helmet, module, outputStack);
         }
     
         @Override
-        public @NotNull Codec<RecipeModuleAddition> codec(){
-            return RecordCodecBuilder.create(
-                codec -> codec.group(
-                ResourceLocation.CODEC.fieldOf("helmet").forGetter(recipe -> BuiltInRegistries.ITEM.getKey(recipe.helmet)),
-                Codec.STRING.fieldOf("module").forGetter(recipe -> recipe.module),
-                ItemStack.CODEC.fieldOf("result").forGetter(recipe -> recipe.outputStack)
-            ).apply(codec, (helmetResourceLocation, module, result) -> new RecipeModuleAddition(BuiltInRegistries.ITEM.get(helmetResourceLocation), module, result)));
+        public @NotNull MapCodec<RecipeModuleAddition> codec(){
+            return CODEC;
         }
-    
+
         @Override
-        public void toNetwork(FriendlyByteBuf buffer, RecipeModuleAddition recipe) {
+        public StreamCodec<RegistryFriendlyByteBuf, RecipeModuleAddition> streamCodec() {
+            return STREAM_CODEC;
+        }
+
+        public static void toNetwork(RegistryFriendlyByteBuf buffer, RecipeModuleAddition recipe) {
             buffer.writeResourceLocation(BuiltInRegistries.ITEM.getKey(recipe.helmet));
             buffer.writeUtf(recipe.module);
-            buffer.writeItem(recipe.outputStack);
+            ItemStack.STREAM_CODEC.encode(buffer, recipe.outputStack);
         }
     }